requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
reportlab==4.0.8
xhtml2pdf==0.2.16
openpyxl==3.1.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
requests==2.31.0
django-filter==23.4
django-extensions==3.2.3
django-widget-tweaks==1.5.0
whitenoise==6.6.0
gunicorn==21.2.0
pandas==2.3.1
dj-database-url==2.1.0
